Fix toSnakeCase bug when using camelCase strings
The regex that was previously used to convert strings to snake_case did
not account for camelCase or StartCamelCase strings, so it simply
downcased the entire string and returned that.

This adds a new regex that is used to handle camelCase and StartCamelCase
strings, as well as the strings previously present in the test cases.

This also adds a new function called `splitByNonDowncaseAlphaNumeric`
that will split a string based on this new regex, trim whitespace from
all strings, and then reject empty strings.

Fixes #169